requirements.txt
filecache==0.81
icecream==2.1.3
pytz==2022.4
nordpool==0.3.3
django==4.1.2
requests==2.28.1
python-dateutil==2.8.2